pyproject.toml
[project]
name = "product-intel-agent"
version = "0.1.0"
description = "A multi-agent system for competitive intelligence and market monitoring."
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "google-adk>=0.1.0",
    "google-genai>=1.0.0",
    "mcp>=1.0.0",
    "fastmcp>=2.13.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "google-cloud-logging>=3.0.0",
    "google-cloud-aiplatform>=1.38.0",
    "google-cloud-secret-manager>=2.26.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
]
